Removed remaining project related functions in WorkbenchDataManager IBP-5699
